E-mail Client Code
Run after database has been set up and data has been added.  Runs in
real-time - if data is not being added while this is running, it will
report zeroes.  Equipment failed while correcting a issue where an error
is thrown when the data is over budget as the first iteration of code.